pyproject.toml
[tool.poetry]
name = "promptlens"
version = "0.1.0"
description = "Lightweight LLM evaluation tool for prompts, agents, and workflows"
authors = ["sparker <sparker@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/sparker/promptlens"
repository = "https://github.com/sparker/promptlens"
keywords = ["llm", "evaluation", "testing", "ai", "prompts", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
anthropic = "^0.40.0"
openai = "^1.54.0"
google-generativeai = "^0.8.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
click = "^8.1.0"
rich = "^13.7.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.9"
mypy = "^1.8.0"

[tool.poetry.scripts]
promptlens = "promptlens.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=promptlens --cov-report=html --cov-report=term"
